fixed #375 fixed #429
