+++ tests/fake_webapp.py
@@ -4,7 +4,7 @@
 # Use of this source code is governed by a BSD-style
 # license that can be found in the LICENSE file.
 
-from flask import Flask, request, abort, Response
+from flask import Flask, request, abort, Response, redirect, url_for
 from os import path
 from functools import wraps
 
@@ -128,6 +128,17 @@ def popup():
 @requires_auth
 def auth_required():
     return "Success!"
+
+
+@app.route('/redirected')
+def redirected():
+    location = '{}?{}'.format(url_for('redirect_location'), 'come=get&some=true')
+    return redirect(location)
+
+
+@app.route('/redirect-location')
+def redirect_location():
+    return EXAMPLE_REDIRECT_LOCATION_HTML
 
 
 def start_flask_app(host, port):
